clean code

+++ controllers/favorites.controller.js
@@ -1,6 +1,7 @@
 import prisma from "../utils/db.js";
+import AppError from "../utils/AppError.js";
 
-const addFavorite = async (req, res) => {
+const addFavorite = async (req, res, next) => {
   const { carId } = req.body;
   try {
     await prisma.favorite.create({
@@ -10,14 +11,13 @@ const addFavorite = async (req, res) => {
       },
     });
 
-    res.status(201).json("Add to favorites successfully!");
+    res.status(201).json("Added to favorites successfully!");
   } catch (error) {
-    res
-      .status(500)
-      .json({ error: "Failed to add to favorites.", error: error.message });
+    next(new AppError("Failed to add to favorites.", 500, error.message));
   }
 };
-const getFavorites = async (req, res) => {
+
+const getFavorites = async (req, res, next) => {
   try {
     const favorites = await prisma.favorite.findMany({
       where: { buyerId: req.userId },
@@ -40,26 +40,28 @@ const getFavorites = async (req, res) => {
       favorites,
     });
   } catch (error) {
-    res
-      .status(500)
-      .json({ error: "Failed to fetch favorites.", error: error.message });
+    next(new AppError("Failed to fetch favorites.", 500, error.message));
   }
 };
 
-const removeFavorite = async (req, res) => {
+const removeFavorite = async (req, res, next) => {
   try {
     const { id } = req.params;
+    const favorite = await prisma.favorite.findUnique({
+      where: { id: parseInt(id) },
+    });
+
+    if (!favorite) {
+      return next(new AppError("Favorite not found.", 404));
+    }
+
     await prisma.favorite.delete({
       where: { id: parseInt(id) },
     });
     res.status(204).json({ message: "Favorite removed." });
   } catch (error) {
-    res
-      .status(500)
-      .json({
-        error: "Failed to remove from favorites.",
-        error: error.message,
-      });
+    next(new AppError("Failed to remove from favorites.", 500, error.message));
   }
 };
+
 export { addFavorite, getFavorites, removeFavorite };
